Fix for a possible future bug
Before if a new file was added in the git repo it would fail because it would not be able to remove it from the users program due too it not existing.

+++ Scripts/files.py
@@ -349,11 +349,12 @@ def check_for_update(currentVersion, updateReleaseChannel, silentCheck=False):
             extraFolderPath = f"{cwd}/{stagingFolder}/{extraFolderPath[0]}"
             
             for file_name in os.listdir(extraFolderPath):
-              try:
-                  os.remove(file_name)
-              except IsADirectoryError:
-                  rmtree(file_name)
-              move(f"{extraFolderPath}/{file_name}", f"{cwd}/{file_name}")
+              if os.path.exists(file_name):
+                try:
+                    os.remove(file_name)
+                except IsADirectoryError:
+                    rmtree(file_name)
+                move(f"{extraFolderPath}/{file_name}", f"{cwd}/{file_name}")
 
           rmtree(stagingFolder)
           print(f"\n> Update completed: {currentVersion} ==> {F.GREEN}{latestVersion}{S.R}")
